Fix major bugs - README updated - add Binance features graph
